fix(daemon): unify SpawnTTY and childreap contracts across platforms, close PTY resize channels, and remove dead Windows shell helpers
