+++ apps/daemon/pkg/common/spawn_tty_linux.go
@@ -6,7 +6,6 @@
 package common
 
 import (
-	"context"
 	"fmt"
 	"io"
 	"os"
@@ -16,24 +15,6 @@ import (
 
 	"github.com/creack/pty"
 )
-
-type TTYSize struct {
-	Height int
-	Width  int
-}
-
-type SpawnTTYOptions struct {
-	// Ctx, when non-nil, bounds the spawned session: on cancellation the
-	// TTY is torn down and the attached process is terminated. Not yet
-	// honored by the Linux implementation.
-	Ctx    context.Context
-	Dir    string
-	StdIn  io.Reader
-	StdOut io.Writer
-	Term   string
-	Env    []string
-	SizeCh <-chan TTYSize
-}
 
 func SpawnTTY(opts SpawnTTYOptions) error {
 	shell := GetShell()
@@ -46,7 +27,13 @@ func SpawnTTY(opts SpawnTTYOptions) error {
 	cmd.Env = append(cmd.Env, fmt.Sprintf("SHELL=%s", shell))
 	cmd.Env = append(cmd.Env, opts.Env...)
 
-	f, err := pty.Start(cmd)
+	var f *os.File
+	var err error
+	if opts.InitCols >= 1 && opts.InitRows >= 1 {
+		f, err = pty.StartWithSize(cmd, &pty.Winsize{Rows: uint16(opts.InitRows), Cols: uint16(opts.InitCols)})
+	} else {
+		f, err = pty.Start(cmd)
+	}
 	if err != nil {
 		return err
 	}
