+++ apps/daemon/pkg/common/spawn_tty_windows.go
@@ -9,33 +9,10 @@ import (
 	"context"
 	"io"
 	"log/slog"
+	"os"
 
 	"github.com/UserExistsError/conpty"
 )
-
-const (
-	defaultTTYCols = 80
-	defaultTTYRows = 24
-)
-
-type TTYSize struct {
-	Height int
-	Width  int
-}
-
-type SpawnTTYOptions struct {
-	// Ctx, when non-nil, bounds the spawned session: on cancellation the
-	// TTY is torn down and the attached process is terminated.
-	Ctx      context.Context
-	Dir      string
-	StdIn    io.Reader
-	StdOut   io.Writer
-	Term     string
-	Env      []string
-	InitCols int
-	InitRows int
-	SizeCh   <-chan TTYSize
-}
 
 func SpawnTTY(opts SpawnTTYOptions) error {
 	shell := GetShell()
@@ -60,6 +37,12 @@ func SpawnTTY(opts SpawnTTYOptions) error {
 	if opts.Dir != "" {
 		cptyOpts = append(cptyOpts, conpty.ConPtyWorkDir(opts.Dir))
 	}
+	if len(opts.Env) > 0 {
+		// ConPtyEnv replaces the child's entire environment block, so
+		// append the extras to the inherited environment to keep the
+		// Linux semantics of Env.
+		cptyOpts = append(cptyOpts, conpty.ConPtyEnv(append(os.Environ(), opts.Env...)))
+	}
 
 	cpty, err := conpty.Start(cmdLine, cptyOpts...)
 	if err != nil {
@@ -67,16 +50,6 @@ func SpawnTTY(opts SpawnTTYOptions) error {
 		return err
 	}
 	defer cpty.Close()
-
-	// Bind the session lifetime to the caller's context so a disconnected
-	// client unblocks the Wait below and the deferred Close kills the
-	// attached process instead of leaking it.
-	parent := opts.Ctx
-	if parent == nil {
-		parent = context.Background()
-	}
-	ctx, cancel := context.WithCancel(parent)
-	defer cancel()
 
 	go func() {
 		for win := range opts.SizeCh {
@@ -87,8 +60,6 @@ func SpawnTTY(opts SpawnTTYOptions) error {
 	}()
 
 	go func() {
-		// Client EOF/disconnect ends the stdin copy: cancel so Wait returns.
-		defer cancel()
 		if _, err := io.Copy(cpty, opts.StdIn); err != nil && err != io.EOF {
 			slog.Debug("ConPTY stdin copy error", "error", err)
 		}
@@ -100,15 +71,8 @@ func SpawnTTY(opts SpawnTTYOptions) error {
 		}
 	}()
 
-	exitCode, err := cpty.Wait(ctx)
+	exitCode, err := cpty.Wait(context.Background())
 	if err != nil {
-		if ctx.Err() != nil {
-			// Cancellation is a normal session end: the caller's ctx fired
-			// (client connection gone) or stdin hit EOF. The deferred Close
-			// tears down the ConPTY and terminates the attached process.
-			slog.Debug("ConPTY session ended by client disconnect")
-			return nil
-		}
 		slog.Debug("ConPTY wait error", "error", err)
 		return err
 	}
